LPS-56048 Move StagedModelDataHandlers into module
