+++ modules/apps/message-boards/message-boards-test/test/integration/com/liferay/message/boards/lar/test/dependencies/attachment.txt
